Está a funcionar acho eu

+++ fich_c/server_default.c
@@ -21,10 +21,9 @@
 #include "../fich_h/client_default.h"
 /*------------------------------*/
 
-int conta_users, *users, user_to_kill, *contaClientes, *ocupantesL;
-char **users_nome, nome_to_kill[9], **matriz, **matrizP;
+int conta_users = 0, *users, *contaClientes, *ocupantesL;
+char **users_nome, **matriz, **matrizP;
 static server server1;
-static pthread_mutex_t *lock;
 
 //ocupantesL está iniciada a -1
 //existem tantas lock como linhas (uma por cada)
@@ -39,52 +38,13 @@ void inicializaContadorClientes(){
 
 	contaClientes= malloc(server1.MEDIT_NUM_PIPES*sizeof(int));
 
-}
+	for(int i = 0; i < server1.MEDIT_NUM_PIPES; i++)
+	{
 
-void sig_handler2(int signo)
-{
-	int encontrou, i, e;
-	encontrou=i=0;
-    	if (signo == SIGUSR2){        
-        for(i=0; i<conta_users;i++){
-			if(users[i]==user_to_kill){
-				if(i!=conta_users-1){
-					users[i]=users[i+1];
-				}
-				encontrou++;
-			}
-			else{
-				if(encontrou>0)
-					if(i!=conta_users-1)
-						users[i]=users[i+1];
-			}
-		}
+		contaClientes[i] = 0;
 
-		encontrou = 0;
-		
-		for(i = 0; i < conta_users; i++)
-		{
-			
-			if (strcmp(nome_to_kill, users_nome[i]) == 0) 
-			{
-				if(i != conta_users - 1)
-					strcpy(users_nome[i], users_nome[i + 1]);
-				encontrou++;
-			}
-			else
-			{
+	}
 
-				if(encontrou > 0)
-					if(i != conta_users - 1)
-						strcpy(users_nome[i], users_nome[i + 1]);
-			}
-			
-		}
-		
-		conta_users--;
-
-		users = (int*)realloc(users, conta_users*sizeof(int));
-    }
 }
 
 void pipe_ini(int *myFifo, char *nomePipe){
@@ -93,17 +53,43 @@ void pipe_ini(int *myFifo, char *nomePipe){
 
 	if(*myFifo==-1){
 		if(errno!=EEXIST){
-				fprintf(stderr, "\nErro no mkfifo()\n");
-				exit(-1);
+			fprintf(stderr, "\nErro no mkfifo()\n");
+			exit(-1);
  		}
 	}
 }
 
-int verifica_user(char *nomeFicheiro, char *username, char *exe, char *nPipe){
+void distribui_user_P(char *nPipe)
+{
+	int men, posMen;
+	
+	posMen = 0;
+	men = contaClientes[0];
+
+	for(int i = 1; i < server1.MEDIT_NUM_PIPES; i++){
+
+		if(men < contaClientes[i])
+		{
+
+			men = contaClientes[i];
+			posMen = i;
+
+		}
+
+	}
+
+	sprintf(nPipe,"%d", posMen);
+}
+
+int verifica_user(char *nomeFicheiro, int pid, char *username, char *exe, char *nPipe){
 
 	FILE *f;
 	char auxiliar[9];
 	int i, aux, l, minimo;
+
+	if (conta_users == server1.MEDIT_MAXUSERS) {	//Já não pode entrar mais utilizadores
+		return 0;
+	}
 
 	if((f=fopen(nomeFicheiro,"r"))==NULL){
 
@@ -113,67 +99,80 @@ int verifica_user(char *nomeFicheiro, char *username, char *exe, char *nPipe){
 	}
 
 	while(fscanf(f,"%s",auxiliar)==1){
-		if(strcmp(auxiliar, username)==0){
-			if(conta_users > 1)
-				for(i = 0; i < conta_users; i++)
+		
+		if (strcmp(auxiliar, username)==0) {
+			
+			if (conta_users > 0) {
+				
+				for(i = 0; i < conta_users -1; i++)
 				{
-					if(strcmp(users_nome[i],username)==0){
-						conta_users--;
-						users = (int *) realloc(users,conta_users * sizeof(int));
+
+					if(strcmp(auxiliar, matriz[i])==0) //Verifica se existe já na matriz de utilizadores
+					{
+
 						fclose(f);
 						return 0;
+
 					}
-						
-				}			
 
-			if(conta_users > server1.MEDIT_MAXUSERS)				
-				return 0;	
-
-			if(conta_users == 1)
-			{				
-				
-				users_nome = (char **) malloc(server1.MEDIT_MAXUSERS * sizeof(char *));
-
-				for(i = 0; i < server1.MEDIT_MAXUSERS; i++)
-					users_nome[i] = (char *) malloc(9 * sizeof(char)); 
-
-				for(i=0; i < 8; i++)
-					users_nome[0][i]=username[i];				
-
-				users_nome[0][8] = '\0';
-
-				contaClientes[0]=1;
-				sprintf(nPipe, "%d", 0);
-
-			}
-			else
-				if(conta_users > 1){			
-
-					for(i=0; i < 8; i++)
-						users_nome[conta_users - 1][i]=username[i];				
-
-					users_nome[conta_users - 1][8] = '\0';
-			
-					aux=0;
-					minimo=contaClientes[0];
-					for(l=1; l<=server1.MEDIT_NUM_PIPES; l++){
-						if(contaClientes[l]<contaClientes[l-1] && contaClientes[l]<minimo){
-							aux=l;
-							minimo=contaClientes[l];
-						}
-					}
-					sprintf(nPipe, "%d", aux);
 				}
 
-			return 1;
+				for(i=0; i<server1.MEDIT_MAXUSERS; i++)
+				{
 
-		}	
+					if(users[i] == -1)
+					{
 
+						distribui_user_P(nPipe);
+						strcpy(users_nome[i], username);
+						users[conta_users] = pid;
+						conta_users++;		
+										
+					}
+
+				}
+				
+				fclose(f);
+				return 1;
+
+			}			
+			else	//Se não houver nenhum utilizador
+			{
+
+				distribui_user_P(nPipe);
+				strcpy(users_nome[0],username);
+				users[0] =  pid;
+				conta_users++;	
+				fclose(f);
+				return 1;
+
+			}			
+			
+		}
+		
 	}
-
 	fclose(f);
 	return 0;
 
+}
+
+void desliga_user(int pid)
+{
+	int i;
+
+	for(i = 0; i < server1.MEDIT_MAXUSERS; i++)
+	{
+
+		if(users[i] == pid)
+		{
+
+			conta_users--;
+			users[i] = -1;
+			strcpy(users_nome[i], " ");
+
+		}
+
+	}
 }
 
 void limpa(){
@@ -197,7 +196,8 @@ void kill_thread(){
 	int i;
 
 	for(i=0; i < conta_users; i++){
-		kill(users[i], SIGUSR1);
+		if(users[i] != -1)
+			kill(users[i], SIGUSR1);
 	}
 	
 	pthread_exit(0);
@@ -211,55 +211,42 @@ void * le_pipe (void * arg){
 	servCli resposta;
 
 	signal(SIGUSR1, kill_thread);
-	signal(SIGUSR2, sig_handler2);
 
-	if( (fd=open(MEDIT_NAME_PIPE_PRINCI_V, O_RDWR))==-1){
+	if( (fd=open(server1.MEDIT_NAME_PIPE_PRINCI, O_RDWR))==-1){
 		fprintf(stderr, "\nErro ao abir a pipe de leitura\n");
 	}
 	
 	while((nr = read(fd, &recebe, sizeof(cliServ)))>0){
-		//Fazer verificações para a alocação de memória
-		user_to_kill=recebe.pid;
-		
-		for(i = 0; i < conta_users; i++)
-			if(user_to_kill = users[i])
-				strcpy(nome_to_kill, users_nome[i]);		
-		
-		if(conta_users==0){
-			users = (int *) malloc(sizeof(int));
-			conta_users++;
-			users[0]=recebe.pid;
-		}
-		else{
-			for(i=0;i<conta_users;i++){
-				if(users[i]==recebe.pid){
-					conta_vistos++;
-					i=conta_users;
-				}
-			}
-			if(conta_vistos){
-				conta_vistos=0;
-			}
-			else{
-				conta_users++;
-				users = (int *) realloc(users, conta_users * sizeof(int));
-				users[conta_users-1]=recebe.pid;
-			}
-		}
 
-		resposta.estado=0;
-		sprintf(pipePID, "%d", recebe.pid);
-		if( (fd_abrirE=open(pipePID, O_WRONLY))<0){
-			fprintf(stderr, "Erro ao abir a pipe cliente\n");
-			break;
+		if(recebe.estado == 0)
+		{
+
+			resposta.estado=0;
+			sprintf(pipePID, "%d", recebe.pid);
+			if( (fd_abrirE=open(pipePID, O_WRONLY))<0){
+				fprintf(stderr, "Erro ao abir a pipe cliente\n");
+				break;
+			}
+			resposta.pid=getpid();
+			resposta.muda=1;
+			resposta.valID=verifica_user(server1.MEDIT_FICHEIRO, recebe.pid,recebe.nome, "client", resposta.fifo_serv);
+			//vai devolver -1 se não abrir o ficheiro
+			//vai devolver 1 no caso de dar certo 
+			//vai devolver 0 no caso de dar errado
+			nw = write(fd_abrirE, &resposta, sizeof(servCli));
+
 		}
-		resposta.pid=getpid();
-		resposta.muda=1;
-		resposta.valID=verifica_user(server1.MEDIT_FICHEIRO, recebe.nome, "client", resposta.fifo_serv);
-		//vai devolver -1 se não abrir o ficheiro
-		//vai devolver 1 no caso de dar certo 
-		//vai devolver 0 no caso de dar errado
-		nw = write(fd_abrirE, &resposta, sizeof(servCli));
+		else
+			if(recebe.estado == 2)
+			{
+
+				desliga_user(recebe.pid);
+				//Sair do servidor
+					//Tirar o pid
+					//Tirar o nome 
+
+			}
+
 	}	
 }
 
@@ -323,7 +310,6 @@ void * le_pipe1 (void * arg){
     	cliServ recebe;
 
 	signal(SIGUSR1, kill_thread);
-	signal(SIGUSR2, sig_handler2);
 
 
 	while((nr = read(fd, &recebe, sizeof(cliServ)))>0){
@@ -366,13 +352,17 @@ void * le_pipe1 (void * arg){
 		
 
 		if(ocupantesL[recebe.linha] == recebe.pid)
-		{			
+		{		
+
 			if(recebe.caracter=='\n')
 			{			
-
+				
 				for(i = 0; i < server1.MEDIT_MAXCOLUMNS; i++)
 				{
-					
+					while(matriz[recebe.linha][i] == ' ' || matriz[recebe.linha][i] == '\n'){
+						i++;
+					}
+
 					if(matriz[recebe.linha][i] == ' ' || matriz[recebe.linha][i] == '\n')
 					{
 						palavra[j]='\0';
@@ -407,6 +397,7 @@ void * le_pipe1 (void * arg){
 			}
 			else	//se for letra ou caracter especial vai para a posição da matriz
 			{
+				
 				//Fazer validações da matriz
 
 				//caso seja del
@@ -562,8 +553,19 @@ void inicio_matriz()
 {
 
 	int i, j;
-	
-	lock = (pthread_mutex_t *) malloc(server1.MEDIT_MAXLINES * sizeof(pthread_mutex_t)); 
+
+	users = (int *) malloc(server1.MEDIT_MAXUSERS * sizeof(int));
+
+	users_nome = (char **) malloc(server1.MEDIT_MAXUSERS * sizeof(char *));
+
+	for(i = 0; i < server1.MEDIT_MAXUSERS; i++)
+	{
+		users_nome[i] = (char *) malloc(9 * sizeof(char));
+		strcpy(users_nome[i], " ");
+
+		users[i] = -1;
+	}
+
 	ocupantesL = (int *) malloc(server1.MEDIT_MAXLINES * sizeof(int));
 
 	matriz = (char **) malloc(server1.MEDIT_MAXLINES * sizeof(char *));
@@ -593,7 +595,6 @@ void inicio_matriz()
 			matrizP[i][j] = ' ';
 		}
 		ocupantesL[i] = -1;
-		pthread_mutex_init(&lock[i], NULL);
 	}
 		
 }
@@ -610,4 +611,7 @@ void liberta_matriz()
 void liberta_users()
 {
 	free(users);
+	for(int i = 0; i < server1.MEDIT_MAXUSERS; i++)
+    	free(users_nome[i]);
+	free(users_nome);
 }
